requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
Django==4.2.7
djangorestframework==3.14.0
flake8==7.1.1
h11==0.14.0
idna==3.10
mccabe==0.7.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PySocks==1.7.1
pytz==2024.2
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
